requirements.txt
pypdf
scikit-learn
streamlit
python-dotenv
google-generativeai
pillow
numpy
